refactor(bridge): move event schemas into an event submodule
